Entry08 : Second round of finishing with .h and all prototypes ok

+++ Libft/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: aadoue <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 14:00:13 by aadoue            #+#    #+#             */
-/*   Updated: 2022/10/13 14:06:59 by aadoue           ###   ########.fr       */
+/*   Updated: 2022/10/31 15:02:24 by aadoue           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
